+++ examples/CMakeLists.txt
@@ -1,5 +1,4 @@
-# add_subdirectory(lvr2_channel_usage)
-add_subdirectory(lvr2_raycaster)
-# add_subdirectory(lvr2_coordinates)
-# add_subdirectory(lvr2_io_features)
-# add_subdirectory(lvr2_scan_io)
+add_subdirectory(channels)
+add_subdirectory(coordinates)
+add_subdirectory(raycasting)
+
