+++ examples/channels/Main.cpp
