--- examples/lvr2_io_features/types/Hyper.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-struct Hyper {
-
-};
