--- examples/lvr2_io_features/types/Image.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-struct Image {
-
-};
